requirements.txt
pypdf>=4.0.0
python-docx>=1.0.0
openai>=1.40.0
tqdm>=4.66.0
dashscope>=1.17.0
